Reduce redundancy across CLI, workflow, and templates
